Folders restructuring
