+++ src/UserContext.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, useEffect, useState } from 'react';
+import { createContext, useContext, useState } from 'react';
 import getUser, { IUser } from '~/services/getUser';
 
 interface IUserContext {
