+++ src/shared/components/Header/header-style.scss
@@ -1,4 +1,4 @@
-@import '../../../../style/_variables';
+@import '../../../style/_variables';
 
 .header {
   height: 160px;
