+++ src/shared/components/Header/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Header';
