+++ src/shared/components/LoadingDots/loading-dots.scss
