+++ src/shared/components/PrivateRoute/PrivateRoute.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { Route, RouteProps, useHistory } from 'react-router-dom';
 import {Routes} from '~/constants';
-import useIsAuthenticated from '~/views/PasswordHealth/components/PrivateRoute/useIsAuthenticated';
+import useIsAuthenticated from '~/shared/hooks/useIsAuthenticated';
 
 const PrivateRoute: FC<RouteProps> = ({
   path,
