+++ src/shared/components/PublicRoute/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PublicRoute';
