+++ src/views/ErrorBlock/ErrorBlock.tsx
@@ -1,4 +1,4 @@
-import { FC, memo } from 'react';
+import React, { FC, memo } from 'react';
 
 import './error-style.scss';
