+++ src/views/LoadingScreen/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LoadingScreen';
