+++ src/views/PasswordHealth/components/Filter/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Filter';
+
