+++ src/views/PasswordHealth/components/List/index.ts
@@ -1,0 +1,2 @@
+export { default } from './List';
+
